scan.sh
This will prompt the user to enter the target IP address. Once the user enters the target IP address, the script will run nmap with the specified target IP address and print the nmap output to the console.

You can modify this code to create a more complex CLI interface for nmap. For example, you could add command-line arguments to allow the user to specify the nmap scan type, the port range to scan, and other options. You could also add additional functionality, such as the ability to save the nmap output to a file or to display the nmap output in a more user-friendly format.