sqlinjection.c
This code is just a starting point. You can modify it to add additional checks or to make it more sophisticated. For example, you could use a regular expression to check for more complex SQL injection patterns. You could also add support for checking for SQL injection in other types of input, such as POST data or request headers.