Jump
This is a jump game where the player(user) can move left and right and jump. The player can jump by pressing the space bar and the player will fall off the screen if they reach the bottom. The player can also reach the end of the game by reaching the right side of the screen.